# src/apps/web/public/anss/openapi/saas.openapi.yaml
openapi: "3.1.0"
info:
  title: "OWorker SaaS Starter API"
  version: "0.1.0"
  description: "Generated from src/apps/api/src/capabilities.mjs."
servers:
  - url: "http://localhost:3001"
paths:
...
  /billing/checkout:
    post:
      operationId: "saas_billing_checkout_create"
      summary: "Create a billing checkout redirect contract."
      x-anss-capability-id: "saas.billing.checkout.create"
      x-anss-safety:
        access: "authenticated-write"
        writes: true
        requiresUserConfirmation: true
      x-anss-input-schema:
        type: "object"
        additionalProperties: false
        required:

        properties:
          accountId:
            type: "string"
            description: "Account id for checkout."
          email:
            type: "string"
            description: "Optional customer email."
      x-anss-output-schema:
        type: "object"
        additionalProperties: true
        required:
          - "provider"
        properties:
          provider:
            type: "string"
            description: "Billing provider."
          url:
            type: "string"
            description: "Checkout redirect URL, when created."
          accountId:
            type: "string"
            description: "Account id associated with checkout."
          status:
            type: "string"
            description: "Provider status when checkout could not be created."
          detail:
            type: "string"
            description: "Provider error detail."
      responses:
        200:
          description: "JSON response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: true
                required:
                  - "provider"
                properties:
                  provider:
                    type: "string"
                    description: "Billing provider."
                  url:
                    type: "string"
                    description: "Checkout redirect URL, when created."
                  accountId:
                    type: "string"
                    description: "Account id associated with checkout."
                  status:
                    type: "string"
                    description: "Provider status when checkout could not be created."
                  detail:
                    type: "string"
                    description: "Provider error detail."
        default:
          description: "Error response"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties: true
                required:
                  - "status"
                  - "detail"
                properties:
                  status:
                    type: "string"
                    enum:
                      - "error"
                  detail:
                    type: "string"
                    description: "Human-readable error detail."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: "object"
              additionalProperties: false
              required:

              properties:
                accountId:
                  type: "string"
                  description: "Account id for checkout."
                email:
                  type: "string"
                  description: "Optional customer email."
...